+++ html/mainpage.html
@@ -188,6 +188,6 @@
     <button id="downloadAllRecordsBtn" class="footer-button download-all-button">⬇️ Download All</button>
   </div>
 
-  <script src="../js/mainpage.js"></script>
+  <script type="module" src="../js/mainpage.js"></script>
 </body>
 </html>
